Feat: add verify style
