+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "iSunFA",
-  "version": "0.8.2+33",
+  "version": "0.8.2+34",
   "private": false,
   "scripts": {
     "dev": "next dev",
